Extract state machine types and move character creation to state module
