+++ apps/api/src/app.module.ts
@@ -1,8 +1,9 @@
 ﻿import { Module } from '@nestjs/common';
 
 import { ChatModule } from './modules/chat/chat.module';
+import { StateMachineModule } from './modules/state-machine/state-machine.module';
 
 @Module({
-  imports: [ChatModule],
+  imports: [ChatModule, StateMachineModule],
 })
 export class AppModule {}
