+++ apps/api/src/modules/chat/chat.module.ts
@@ -2,11 +2,9 @@
 
 import { ChatController } from './chat.controller';
 import { ChatService } from './chat.service';
-import { CharacterCreationService } from './character-creation.service';
 
 @Module({
   controllers: [ChatController],
-  providers: [ChatService, CharacterCreationService],
-  exports: [CharacterCreationService],
+  providers: [ChatService],
 })
 export class ChatModule {}
