+++ apps/api/src/modules/state-machine/character-creation.service.spec.ts
@@ -1,4 +1,5 @@
-﻿import { CharacterCreationService, type CharacterDraft, type CharacterField } from './character-creation.service';
+﻿import { CharacterCreationService } from './character-creation.service';
+import type { CharacterDraft, CharacterField } from '@ai-dungeon-master/models';
 
 describe('CharacterCreationService', () => {
   let service: CharacterCreationService;
@@ -63,3 +64,5 @@ describe('CharacterCreationService', () => {
     expect(Object.values(character.abilities)).toContain(15);
   });
 });
+
+
