+++ apps/api/src/modules/state-machine/state-machine.module.ts
@@ -1,0 +1,9 @@
+﻿import { Module } from '@nestjs/common';
+
+import { CharacterCreationService } from './character-creation.service';
+
+@Module({
+  providers: [CharacterCreationService],
+  exports: [CharacterCreationService],
+})
+export class StateMachineModule {}
